+++ packages/analytics-js-common/src/constants/loggerContexts.ts
@@ -11,7 +11,7 @@ const LOAD_CONFIGURATION = 'LoadConfiguration';
 const EVENT_REPOSITORY = 'EventRepository';
 const EXTERNAL_SRC_LOADER = 'ExternalSrcLoader';
 const HTTP_CLIENT = 'HttpClient';
-const RSA = 'RudderStackAnalytics';
+const RS_APP = 'RudderStackApplication';
 const ANALYTICS_CORE = 'AnalyticsCore';
 
 export {
@@ -28,6 +28,6 @@ export {
   EVENT_REPOSITORY,
   EXTERNAL_SRC_LOADER,
   HTTP_CLIENT,
-  RSA,
+  RS_APP,
   ANALYTICS_CORE,
 };