+++ packages/analytics-js-plugins/__tests__/bugsnag/utils.test.ts
@@ -695,6 +695,13 @@ describe('Bugsnag utilities', () => {
         },
         ['key4', 'key6'], // excluded keys
       ],
+      [
+        {
+          someKey: BigInt(123),
+        },
+        undefined,
+        [],
+      ],
     ];
 
     it.each(tcData)('should convert signals to JSON %#', (input, expected, excludes) => {