+++ packages/analytics-js/__tests__/services/HttpClient/HttpClient.test.ts
@@ -230,4 +230,26 @@ describe('HttpClient', () => {
       url: `${dummyDataplaneHost}/emptyJsonSample`,
     });
   });
+
+  it('should handle if input data contains non-stringifiable values', done => {
+    const callback = (response: any) => {
+      expect(response).toBeUndefined();
+      expect(defaultErrorHandler.onError).toHaveBeenCalledTimes(1);
+      expect(defaultErrorHandler.onError).toHaveBeenCalledWith(
+        new Error('Failed to prepare data for the request.'),
+        'HttpClient',
+      );
+      done();
+    };
+    clientInstance.getAsyncData({
+      callback,
+      url: `${dummyDataplaneHost}/nonStringifiableDataSample`,
+      options: {
+        data: {
+          a: 1,
+          b: BigInt(1),
+        },
+      },
+    });
+  });
 });